I was seeing 40 second load times on a bar chart with roughly 101 rows in IE8. I looked into the highcharts source and realized a lot of unnecessary reflows were making things slow.
For my chart these changes took the render time down from 40 seconds to around 1.9 seconds.

To address this I added 3 functions to the Tick object,

1) attachLabel
2) updateTransformLabel
3) computeBBox

and 3 properties to the Tick.label object,

1) elemWidth
2) elemHeight
3) deferUpdateTransform

and 1 property to the Tick object
1) attached

in the original course of action the each loop in getOffset would call an "addLabel" then immediately do a lookup of a BBox which would force a reflow, not giving the browser a chance to group all the adds together.

My solution to this was separating the lines that do lookups and the lines that do manipulation.

The new each statement first loops over all the tickPositions and addLabel no longer does the .add(gridAxis)

Immediately after doing all the addLabels, another each loop goes through and attaches all the labels but doesn't do the updateTransform, instead deferring.

Another each loop goes through all the attached labels and computes the height and width and stores them in elemWidth and elemHeight (calls to offsetWidth/offsetHeight were causing the reflows)

The next loop goes through all the tickpositions again and finally updateTransforms them all.

Then finally we loop through everything a last time and compute the labelOffset.