Fix handling of HTML entities invalid in XML
The export module relies on `innerHTML` to retrieve the SVG representation
of the current chart. Unfortunately, HTML is **not** XML. Specifically, if
a node contains U+00A0 (NO-BREAK SPACE) or U+00AD (SOFT HYPHEN), some
browsers will return an HTML entity (`&nbsp;` and `&shy;`, respectively)
rather than the actual character. Since these entities are not defined in
XML, the SVG document is invalid and processing it into an image fails.

This change simply replaces all occurrences of `&nbsp;` and `&shy;` with
the correct characters.