Remove need to configure export server location
By using a relative URL of `action="./"` in the JSP, we can remove a
bunch of manual configuration needed to set up the server. The URL of
the '/demo' resource is always known to be at the same level as the POST
submission endpoint, so we can use a relative URL to access it.