Possibility to extend form data for exporting
Add the possibility to add form data in exporting post.
We would like to add some form data because we want to make an excel export with the data and the chart picture. We have to known some other informations to execute a sql query to have the correct data.