Map: add mapNavigation.mouseWheelSensitivity option for zooming
With a typical mouse wheel, 5 to 6 mousewheel events (or magnitude
1 or -1) would be generated by a single full finger scroll. Likewise,
some 5 to 6 wheel events could easily be triggered by a finger dragged
over touchpad devices.

Highmaps zoomed-in to 200% and zoomed-out to 50% per each such event.
It was pretty easy to zoom too much much too fast.

This commit reduces the percent amount to 110% (and 90%) by default,
and introduces a mapNavigation.mouseWheelSensitivity option where
the user can customize it.