New classes for 3D Renderer: Face3D and Polyhedron
The cuboid function is very handy, but not enough for defining complex shapes.

A Face3d applies the projection to it's vertexes and creates a SVG polygon (Which
is only visible if the polygon is oriented CCW, implementing backface culling).

A Polyhedron acts as a collection of Face3D objects

TODO: Support for animation

TODO: Update Cuboid to be a subclass of Polyhedron