[Highcharts-3D] Automatically move Axis labels/titles to the best edge
In a 3D chart, the labels for the Y axis should always be on the left-most available edge (right-most if opposite),
and X and Z axis should always be on the lowet X/Z available edges (uppermost if opposite).

An edge is considered available if only one of it's adjacent faces is visible, or if both are visible, but one of them is facing front and the other is facing back.
If we used any of the eliminated other edges, the result would be cluttered.

Given these constraints, it is possible that for a given projects, there won't be any edge available, and some of the axis labels will be hidden.
E.g., if looking from above (alpha=90, frame.visible='auto'), all labels are hidden. If looking from beloc (alpha=-90), Only X and Z axis labels are visible.
Those are extreme examples -- Usually all 3 labels are visible.