Catch the removeDirectory error as to not break the dist pipeline
When running gulp dist, it cleans out the build/dist by removing it but the highcharts-assembler throws an error if the directory you are removing is not there. This breaks the dist pipeline unless you create an empty build/dist folder. Here, we are catching the error as to not break the pipeline.