Tools: Wait for svg file writes to complete.
There seems to be an issue with karma-imagecapture-reporter exiting before all file writes are complete. At least that is the theory, but it is hard to reproduce. Latest example is https://circleci.com/gh/highcharts/highcharts/25460 where there are no reported file write errors, which indicates that nothing has gone wrong (as it then should produce a log line with a short error message).

Thus I have added some code to make sure the files are written before karma is exiting, which is more or less same approach as karma-junit-reporter uses.