Bugfix/23279-styled-mode-solidgauge-stops
Fixed #23279, color CSS classes were removed when neither stops nor
minColor/maxColor were defined.

The stops options are always initialized either with the values provided
through options or based on the `minColor`/`maxColor`. However, when
neither of these value are defined through options, then the stops are
initialized with invalid colors. This adds another check to ensure that
the `highcharts-color-{n}` CSS classes are not removed when the stops
don't contain valid colors.

The tests were added to the same file that was used for the other case,
but it can be refactore to its own suite if so desired.